Revert "Revert "Stop entity-scoped nerdstorage writes""
