+++ internal/install/execution/installevents_reporter_integration_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/newrelic/newrelic-cli/internal/install/types"
 	"github.com/newrelic/newrelic-client-go/v2/newrelic"
 	"github.com/newrelic/newrelic-client-go/v2/pkg/config"
+	"github.com/newrelic/newrelic-client-go/v2/pkg/workloads"
 )
 
 func TestInstallEventsReporter_Basic(t *testing.T) {
@@ -51,4 +52,14 @@ func TestInstallEventsReporter_Basic(t *testing.T) {
 
 	err = r.RecipeInstalled(status, evt)
 	require.NoError(t, err)
+}
+
+func createEntity(t *testing.T, accountID int, c *newrelic.NewRelic) string {
+	i := workloads.WorkloadCreateInput{
+		Name: "testEntity",
+	}
+	e, err := c.Workloads.WorkloadCreate(accountID, i)
+	require.NoError(t, err)
+
+	return string(e.GUID)
 }
